@@ -4,37 +4,32 @@
     "type": "project",
     "license": "proprietary",
     "require": {
-        "php": "^7.4",
+        "php": ">=8.0",
         "ext-mbstring": "*",
-        "cycle/annotated": "^2.0",
-        "cycle/migrations": "^1.0",
-        "cycle/orm": "^1.0",
-        "cycle/proxy-factory": "^1.0",
-        "ramsey/uuid": "^4.1",
-        "spiral/broadcast": "^2.0",
-        "spiral/database": "^2.3",
-        "spiral/dotenv-bridge": "^1.0",
-        "spiral/framework": "^2.6",
-        "spiral/jobs": "^2.0",
-        "spiral/migrations": "^2.0",
-        "spiral/monolog-bridge": "^1.0",
+        "doctrine/collections": "^1.6",
+        "spiral/cycle-bridge": "^1.0",
+        "spiral/framework": "^2.9",
         "spiral/nyholm-bridge": "^1.0",
-        "spiral/prototype": "^1.0",
-        "spiral/roadrunner": "^1.4",
-        "spiral/scaffolder": "^2.0",
-        "spiral/stempler-bridge": "^1.0"
+        "spiral/roadrunner-bridge": "^1.0",
+        "spiral/roadrunner-broadcast": "^2.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^8.5|^9.0"
+        "phpstan/phpstan": "^1.4",
+        "phpunit/phpunit": "^9.5",
+        "spiral/testing": "^1.1"
     },
     "scripts": {
-        "post-install-cmd": [
+        "post-create-project-cmd": [
             "php -r \"copy('.env.sample', '.env');\"",
             "php app.php encrypt:key -m .env",
-            "php app.php configure -vv"
+            "php app.php configure -vv",
+            "rr get-binary"
         ]
     },
     "autoload": {
+        "files": [
+            "app/src/polyfill.php"
+        ],
         "psr-4": {
             "App\\": "app/src"
         }
@@ -44,11 +39,11 @@
             "Tests\\": "tests"
         }
     },
-    "extra": {
-        "publish-cmd": "php app.php publish"
-    },
     "config": {
-        "sort-packages": true
+        "sort-packages": true,
+        "allow-plugins": {
+            "spiral/composer-publish-plugin": false
+        }
     },
     "minimum-stability": "dev",
     "prefer-stable": true
