@@ -4,32 +4,23 @@ declare(strict_types=1);
 
 namespace App\Bootloader;
 
-use App\Database\User;
-use Spiral\Auth\AuthScope;
+use Spiral\Auth\AuthContextInterface;
 use Spiral\Boot\Bootloader\Bootloader;
 use Spiral\Bootloader\Http\WebsocketsBootloader;
 
 class TopicsBootloader extends Bootloader
 {
-
     protected const DEPENDENCIES = [
-        WebsocketsBootloader::class
+        WebsocketsBootloader::class,
     ];
 
     public function boot(WebsocketsBootloader $ws): void
     {
-        $ws->authorizeTopic(
-            'channel',
-            function () {
-                return true;
-            }
-        );
-        $ws->authorizeTopic('channel.{uuid}', function ($uuid, AuthScope $auth): bool {
-            /** @var User $user */
-            $user = $auth->getActor();
-            if ($user !== null && $user->getUuid() === $uuid) {
-                return true;
-            }
+        $ws->authorizeTopic('channel', function (): bool {
+            return true;
+        });
+        $ws->authorizeTopic('channel.{uuid}', function ($uuid, AuthContextInterface $authContext): bool {
+            return $authContext->getActor()?->getUuid() === $uuid;
         });
     }
 }
